#7 Added dto folder
