+++ .github/workflows/main.yaml
@@ -2,7 +2,7 @@ name: pipeline
 on:
   push:
     branches:
-    - main
+      - main
 jobs:
   build:
     runs-on: ubuntu-latest
@@ -11,7 +11,7 @@ jobs:
       - name: Setup Go environment
         uses: actions/setup-go@v3.5.0
         with:
-          go-version: '1.20.0'
+          go-version: "1.20.0"
           architecture: amd64
       - name: Build app
         run: env CGO_ENABLED=0 GOOS="linux" GOARCH="amd64" go build -o main "./src/main.go"
