+++ src/dto/create_user_dto.go
@@ -1,0 +1,7 @@
+package dto
+
+type CreateUserDto struct {
+	Username string   `json:"username"`
+	Password string   `json:"password"`
+	Roles    []string `json:"roles"`
+}
