+++ src/dto/login_user_dto.go
@@ -1,0 +1,6 @@
+package dto
+
+type LoginUserDto struct {
+	Username string `json:"username"`
+	Password string `json:"password"`
+}
